Rework interactive CLI to use PTY
This makes it appear as a TTY and therefore solves issues due to
programs buffering output. This makes the '$stdout.sync = true'
workaround in guessing-game no longer necessary

There's a bit of a weird separation between interaction sessions
that start a shell and sessions that start with a command — Docker
will stop after the command has exited, but a shell will keep going
forever (until stopped manually). This seems unavoidable unless we
stop you from using an interactive session with a command, but the
only real issue is having to use slightly different expectations,
so it seems OK for now